Moved cylinder series into a subfolder.
