+++ js/masters/modules/cylinder.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/CylinderSeries.js';
+import '../../Series/Cylinder/CylinderSeries.js';
